Updates and fixes
-Merged patch upstream
-optimized make.sh for a better build process
-added make.config option
-added ability to choose whether to download depot_tools or set the location yourself
-fixed init.sh

+++ build/make.sh
@@ -1,4 +1,13 @@
 #!/bin/bash
+
+function asktocontinue() {
+read -p "$1 ([y]es or [N]o): "
+    case $(echo -e $REPLY | tr '[A-Z]' '[a-z]') in
+        y|yes) echo "yes" ;;
+        *)     echo "no" ;;
+    esac
+}
+
 LOCAL_REPO="$(dirname $(dirname $(readlink -f $0)))"
 #Colors
 source $LOCAL_REPO/build/colors.sh
@@ -10,15 +19,27 @@ ROMName=$(basename $LOCAL_REPO)
 
 echo -e "${Yel}Disabling GYP"
 export GYP_CHROMIUM_NO_ACTION=1
+if [ -f $LOCAL_REPO/build/make.config ]
+then
+. $LOCAL_REPO/build/make.config
+else
 export DEPOT_TOOLS_DIR=$HOME/depot_tools
+fi
 if [ -d "$DEPOT_TOOLS_DIR" ]; then
 
 echo -e "${Yel}Found Depot Tools, adding to path"
 export PATH="$PATH":$HOME/depot_tools
 else
 
-echo -e "${Red}Could not find $DEPOT_TOOLS. Downloading from git. If you have it somewhere else, add it to your home directory."
+echo -e "${Yel}Could not find $DEPOT_TOOLS. If you have it somewhere else, add it to your home directory, or make a 'make.config' file in the build folder."
+
+   if [ "no" == $(asktocontinue "Would you like to clone depot_tools to your home directory?") ]
+   then
+   echo "Please configure the location of your depot tools in $LOCAL_REPO/build/make.config, or just clone to your home directory."
+   exit 0
+   else
 git clone git://codeaurora.org/quic/chrome4sdp/chromium/tools/depot_tools.git $HOME
+   fi
 
 echo -e "${Yel}Adding $DEPOT_TOOLS to path"
 export PATH="$PATH":$HOME/depot_tools
@@ -26,30 +47,29 @@ fi
 
 echo -e "${Yel}Cleaning the source and halogen-ifying it"
 # errors on
-set -e
+
 OUTFOLDER=$(date +%d%m)
+value=$( git log -1 | grep -c "XOS" )
 
 cd $LOCAL_REPO/src
 	rm -rf $LOCAL_REPO/src/out
 	git clean -f -d
-	git reset --hard HEAD~1
+if [ $value -eq "0" ] 
+then
+	git reset --hard HEAD
+fi
+if [ $value -eq "1" ] 
+then
+        git reset --hard HEAD~1
+fi
         git checkout HEAD
-        gclient sync
 	git fetch origin
-	git checkout origin/m55
-	git apply $LOCAL_REPO/build/patches/halogenOS.patch && git add -f $(git status -s | awk '{print $2}') && git commit -m "This is XOS Browser"
+	git checkout  origin/m55
+        gclient sync -f
+	git am -3 < ../build/patches/halogenOS.patch
 
   	. build/android/envsetup.sh
   	gn gen out/$OUTFOLDER --args='target_os="android" ffmpeg_branding="Chrome" proprietary_codecs=true enable_nacl=false is_debug=false is_component_build=false target_cpu="arm"'
-
-
-function asktocontinue() {
-read -p "$1 ([y]es or [N]o): "
-    case $(echo -e $REPLY | tr '[A-Z]' '[a-z]') in
-        y|yes) echo "yes" ;;
-        *)     echo "no" ;;
-    esac
-}
 
 if [ "no" == $(asktocontinue "Would you like to start the build now?") ]
 then
